src/main/java/dev/katiejeanne/foodathome/security/CustomUserDetailsService.java
package dev.katiejeanne.foodathome.security;

import dev.katiejeanne.foodathome.domain.User;
import dev.katiejeanne.foodathome.repositories.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.security.crypto.password.PasswordEncoder;
import org.springframework.stereotype.Service;

import java.util.Collections;

@Service
public class CustomUserDetailsService implements UserDetailsService {


    private final UserRepository userRepository;

    private final PasswordEncoder passwordEncoder;

    @Autowired
    public CustomUserDetailsService(UserRepository userRepository, PasswordEncoder passwordEncoder) {
        this.userRepository = userRepository;
        this.passwordEncoder = passwordEncoder;
    }

    @Override
    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {

        User user = userRepository.findByUsername(username)
                .orElseThrow(() -> new UsernameNotFoundException("Username " + username + " not found"));

        // Get the household id for the user's household
        long householdId = user.getHousehold().getId();

        CustomUserDetails userDetails = new CustomUserDetails(user.getUsername(), user.getPassword(), Collections.singletonList(user.getHouseholdRole()), householdId);

        return userDetails;

    }
}
